requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.6.1
gunicorn==23.0.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
swagger-ui-py==23.9.23
typing_extensions==4.14.0
uritemplate==4.2.0
